✅ adding test structure

+++ src/modules/accounts/useCases/deleteAccount/DeleteAccountUseCase.ts
@@ -1,5 +1,6 @@
 import { inject, injectable } from 'tsyringe';
 
+import { AppError } from '../../../../errors/AppError';
 import { IAccountsRepository } from '../../repositories/IAccountsRepository';
 
 @injectable()
@@ -9,8 +10,16 @@ class DeleteAccountUseCase {
     private accountsRepository: IAccountsRepository
   ) {}
 
-  async execute(id: string): Promise<void> {
-    await this.accountsRepository.delete(id);
+  async execute(id: string): Promise<string> {
+    const accountExists = await this.accountsRepository.findById(id);
+
+    if (!accountExists) {
+      throw new AppError('Account not found.');
+    }
+
+    const response = await this.accountsRepository.delete(id);
+
+    return response;
   }
 }
